Client as facade and added readers and writers
Actually we have a major issue with the current design. We cannot use the
same client instance for writing data points using UDP/IP and read data
using the HTTP client. This is because the `adapter` section is designed
as a big block with 2 responsibility per adapter: reading for data and
write information back to InfluxDB:

```php
$adapter = new UdpAdapter(...);
$client = new Client($adapter);
```

Because the Udp/Ip adapter is only `Writeable` we cannot use this
`$client` instance in order to performs also the read operation, we have
to create another client that uses the HTTP adapter and we will force
users to have 2 clients in the application.

This feature allows the final user to mount a `Writer` and a `Reader` in
order to perform completely separate operations and select which
writer implementation is the best one for the current user story and
the same applies to the reader part:

```php
$writer = new Udp\Writer(...);
$reader = new Http\Reader(...);

$client = new Client($reader, $writer);
```

Now the `$client` instance performs both operation: `mark` and `query`
using the Udp\Ip adapter for write information and the HTTP adapter for
read back data.

The drawback that is the library composition now is a bit harder (again)
because of this composition.

This step could be also interesting in order to refactor the actual
**Test Suite** that is effectively a bit pointless and obfuscate
different library goals.