Support for custom query manager
Instead of mapping all database specific queries like: `createDatabase`
or `getDatabases` directly in the `InfluxDB\Client` class we can prepare
a manager that will enable any user to create custom queries and release
them out. In addition a preset of queries like: `'SHOW DATABASES'` and
any ordinary query set is included as part of this project.

The manager uses a simple way to add new queries:

```php
$manager = new Manager($client);
$manager->addQuery("myQuery", function($arg1, $arg2) {
    return sprintf("THIS IS A CUSTOM QUERY WITH: %s %s", $arg1, $arg2);
});

//Use it
$manager->myQuery("test", "ok");
```

Thanks to PHP we can also use objects in order to create a valid
commands in a more concise way:

```php
class MyCommand
{
    public function __toString()
    {
        return "myCommand";
    }

    public function __invoke()
    {
        return "HERE IS MY COMMAND";
    }
}

$manager->addQuery(new MyCommand());

$data = $manager->myCommand();
```

This features release a preset of queries (not engaged by default
anymore):

```php
$manager->addQuery(new CreateDatabase()); //create database command
$manager->addQuery(new DeleteDatabase()); //drop database command
$manager->addQuery(new GetDatabases());   //list existing databases
```